agents: promote the app bridge and tool-loop driver out of evals
The App Builder bridge moves verbatim from evals/surfaces/app.ts to
app-build/bridge.ts, so the eval suite and the app-build Author stage
drive one implementation of the ui_app_* contract instead of two. It
gains two things the browser has no counterpart for, both opt-in:
ui_app_finish, which lets an authoring model end its turn deliberately,
and document()/loadDocument(), so a repair round resumes from the
document the last round produced instead of an empty one.

The provider-driving loop moves out of tool-loop-eval.ts into
app-build/tool-loop.ts: generateLoop over HeadlessTool[], turn cap, call
transcript. The eval keeps its scoring and reads the transcript back.
